+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Released on ?.
 
 - Updated to use the AngleSharp.Css gradient model (#10)
+- Improved rendering of overflow with ellipsis (#11)
 - Added support for CSS grid track sizing (#7)
 
 # 0.4.0